# api/openapi.yaml
openapi: 3.0.3
info:
  title: Indurex Agentic Alert Investigation Engine API
  version: "1.0.0"
  license:
    name: UNLICENSED
  description: |
    API for defining versioned investigation playbooks, ingesting security
    alerts, and running structured, fully-auditable investigations against
    them. A human analyst UI and an automated agent use exactly this API —
    there is no separate "agent" surface.

    The server owns workflow correctness: clients select an `edge_id` at the
    current node; the server determines the resulting node. See
    docs/domain-model.md for the full invariant list.
servers:
  - url: http://localhost:8080
    description: Local docker-compose environment.

# Authentication/authorization is intentionally out of scope for this
# take-home; see README "Production considerations" for the intended
# scheme (designer / analyst / agent / auditor roles).
security: []
...
paths:
...
  /api/v1/investigations/{investigationId}/decisions:
    parameters:
      - $ref: "#/components/parameters/InvestigationId"
    post:
      tags: [investigations]
      operationId: submitDecision
      summary: Submit a decision, advancing the investigation along the selected edge.
      description: |
        The server validates that the investigation is in progress, that
        `edge_id` belongs to the investigation's playbook version, and that
        it originates at the current node — the client can never choose the
        next node directly. Concurrent submissions on the same investigation
        are serialized; exactly one applies.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
          description: |
            Optional client-supplied key, unique per investigation. Retrying
            the same key with an equivalent request body (same `edge_id`,
            `actor`, `rationale`, `evidence`) does not apply the decision
            twice — it is a no-op that returns the current investigation
            state, whatever it is at the time of the retry (not necessarily
            the state right after the original decision, if later decisions
            have since been applied). Retrying the same key with a
            *different* body is rejected with `409 IDEMPOTENCY_KEY_REUSED`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitDecisionRequest" }
            examples:
              knownWorkstation:
                value:
                  edge_id: "5b1f7e2a-2222-4c11-9a4a-000000000001"
                  actor:
                    type: agent
                    id: investigation-agent-v1
                  rationale: "The source address belongs to ENG-WS-14, a registered engineering workstation."
                  evidence:
                    - type: asset_inventory_lookup
                      summary: "10.20.1.44 maps to ENG-WS-14"
                      data:
                        asset: ENG-WS-14
                        trusted: true
      responses:
        "200":
          description: |
            Decision applied. Response is the resulting investigation state:
            either the next decision node with its choices, or the terminal
            resolution.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvestigationState" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Investigation already completed; `edge_id` does not originate at
            the current node / does not belong to this playbook version; or
            `Idempotency-Key` was reused with a different request body
            (`code: IDEMPOTENCY_KEY_REUSED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
...
    InvestigationId:
      name: investigationId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: "playbook version not found"
        details:
          type: array
          items: { $ref: "#/components/schemas/ErrorDetail" }
      additionalProperties: false
    ErrorDetail:
      type: object
      description: Machine-readable detail; shape varies by `code`.
      additionalProperties: true
...
    PlaybookNodeKind:
      type: string
      enum: [decision, terminal]
    PlaybookNode:
      type: object
      required: [id, kind, title]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/PlaybookNodeKind" }
        title: { type: string }
        description: { type: string }
        terminal_resolution:
          type: string
          nullable: true
          description: Required iff kind = terminal.
        metadata:
          type: object
          nullable: true
          additionalProperties: true
...
    Alert:
      type: object
      required: [id, alert_type, title, occurred_at, created_at]
      properties:
        id: { type: string, format: uuid }
        external_id: { type: string, nullable: true }
        alert_type: { type: string }
        title: { type: string }
        description: { type: string }
        payload:
          type: object
          additionalProperties: true
        occurred_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...
    InvestigationStatus:
      type: string
      enum: [in_progress, completed]
    ActorType:
      type: string
      enum: [human, agent]
    Actor:
      type: object
      required: [type]
      properties:
        type: { $ref: "#/components/schemas/ActorType" }
        id: { type: string, nullable: true }
    EvidenceItem:
      type: object
      required: [type, summary]
      properties:
        type: { type: string }
        summary: { type: string }
        data:
          type: object
          additionalProperties: true
    AvailableChoice:
      description: An edge selectable from the current node.
      type: object
      required: [edge_id, label, to_node_id]
      properties:
        edge_id: { type: string, format: uuid }
        label: { type: string }
        description: { type: string, nullable: true }
        to_node_id: { type: string, format: uuid }
    InvestigationStep:
      description: |
        Every step is a decision: exactly one edge selected from `node_id`.
        There is no such thing as an evidence-only, non-advancing step (see
        docs/domain-model.md §8) — a `terminal` root produces zero steps
        rather than a step with no edge.
      type: object
      required: [id, sequence_number, node_id, selected_edge_id, actor, created_at]
      properties:
        id: { type: string, format: uuid }
        sequence_number: { type: integer }
        node_id: { type: string, format: uuid }
        selected_edge_id: { type: string, format: uuid }
        actor: { $ref: "#/components/schemas/Actor" }
        rationale: { type: string, nullable: true }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/EvidenceItem" }
        created_at: { type: string, format: date-time }
...
    SubmitDecisionRequest:
      type: object
      required: [edge_id, actor]
      properties:
        edge_id: { type: string, format: uuid }
        actor: { $ref: "#/components/schemas/Actor" }
        rationale: { type: string, nullable: true }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/EvidenceItem" }
    InvestigationState:
      description: |
        What an agent/UI needs to decide its next action, self-contained:
        includes the full alert (not just its ID) so a resumed agent call —
        e.g. an LLM reasoning over gathered evidence — has the original
        alert payload, current question, choices, and history from one
        request, without a follow-up `GET /alerts/{alertId}`.
      type: object
      required: [id, alert, playbook_version_id, status, current_node, started_at]
      properties:
        id: { type: string, format: uuid }
        alert: { $ref: "#/components/schemas/Alert" }
        playbook_version_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/InvestigationStatus" }
        current_node: { $ref: "#/components/schemas/PlaybookNode" }
        available_choices:
          description: Empty when status = completed.
          type: array
          items: { $ref: "#/components/schemas/AvailableChoice" }
        final_resolution: { type: string, nullable: true }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        steps:
          type: array
          items: { $ref: "#/components/schemas/InvestigationStep" }
...